Modified to use packed data, enhance verbose output, increase efficiency of permutation check
Lots of changes here.

I increased the verbose output in permutation_tests.py.  It's now possible to monitor what is happening as the permutation tests are run.

I changed the tests in permutation_tests.py so that when the pass criteria are met, the test ends.  I changed the tests in permutation_tests.py so that when a test has passed, it no longer is computed.  These two changes mean that the permutation tests are *much* faster.

I increased the output of restart.py under verbose.

I altered the row and column frequency selection in restart.py so that the results are saved in case future analysis is desired.

I changed to_data in util90b.py so that packed data is used.  This allows bit sizes of greater than 8 bits, but their use is curtailed because the rest of the programs don't deal well with the larger size.  I did get a 9 bit sample size to pass the restart tests.

I added two new variables to the arguments list in the parser in util90b.py.  They are a start_position and read_amount.  The start_position is where to start reading random data in the file, and the read_amount is how many bytes to read.  This allows a larger file to be processed in chunks, so that restart can still be run on the output.

I modified iid_main.py, noniid_main.py, and restart.py to utilize the start_position and read_amount parameters.