Seed the pseudo random number generator correctly.
Seeding the pseudo random number generator before every shuffle using a value
that changes once per second will mean that the different threads will be using
the same sequence of _random_ values.  This is not intended.

Instead, seed the pseudo random number generator once globally and each thread
will get a different sequence of values.  This means locking to control access
to rand(3) but shuffling is a relatively small amount of the total time, so it
is unlikely to hurt.